services/payment_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.1
python-jose==3.3.0
APScheduler==3.10.4
alembic==1.13.1
